REFACTOR : css and removed comments
