+++ src/App.css
@@ -10,6 +10,14 @@ body{
   color: white;
 }
 
+
+.bgImage{
+  background-image: url("https://picsum.photos/1600/1000?random");
+  background-repeat: no-repeat;
+  width: 100%;
+  height: 100vh;
+}
+
 button{
   cursor: pointer;
 }
@@ -22,8 +30,6 @@ input{
 
 input:focus{
   color: white;
-  /* font-size: 2rem;
-  font-weight: 400; */
 }
 
 .center{
@@ -39,7 +45,7 @@ input:focus{
     justify-content: center;
     font-size: 1050%;
     letter-spacing: -5px;
-    font-weight: 520;
+    font-weight: 600;
     letter-spacing: 1rem;
 }
 
@@ -141,7 +147,7 @@ input:focus{
 
 .weather-input>p{
   font-size: 1rem;
-  height: 1rem !important;
+  height: 1rem ;
 }
 
 .weather-input>button{
