+++ src/components/clock/Clock.jsx
@@ -1,7 +1,7 @@
  
  import { React, useState} from 'react'
  
- export default function Clock() {
+  function Clock() {
      const time = new Date().toLocaleTimeString().match(/\d{2}:\d{2}|[AMP]+/g).join(' ');
      const [currentTime, setCurrentTime] = useState(time)
 
@@ -16,4 +16,6 @@
      </div>
    )
  }
- 
+
+
+ export { Clock }
