+++ src/components/index.js
@@ -1,0 +1,3 @@
+export { Clock } from './clock/Clock'
+export { Weather } from './weather/Weather'
+export { Quote } from './quoteComponent/Quote'
