+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { HomePage } from './homePage/HomePage';
+export { LandingPage } from './landingPage/LandingPage';
+export { MainPage } from './mainPage/MainPage';
